refactor: 🗃️ move 'variables.ts' to 'src/data'
