+++ src/data/variables.ts
@@ -1,0 +1,132 @@
+const DEFAULT_USER: string = 'DefaultPool2';
+const ERROR_FAILED_COLLECTION: string = 'ERROR: Failed to fetch collection';
+const ERROR_TOO_MANY: string = 'ERROR: Too many API calls!';
+const BACKEND_URL: string = 'http://localhost:3001/user/';
+const API_ISSUE_REPORT: string =
+  'https://github.com/Cyggnus/ShuffleGenie-API/issues/new?assignees=&labels=&projects=&template=bug_report.md&title=';
+const MIDDLE_NAME: string[] = [
+  'Silly',
+  'Crazy',
+  'Awesome',
+  'Fantastic',
+  'Glorious',
+  'Magnificent',
+  'Hilarious',
+  'Witty',
+  'Epic',
+  'Brilliant',
+  'Incredible',
+  'Wonderful',
+  'Dazzling',
+  'Majestic',
+  'Marvelous',
+  'Phenomenal',
+  'Spectacular',
+  'Thrilling',
+  'Zany',
+  'Whimsical',
+  'Peculiar',
+  'Bizarre',
+  'Curious',
+  'Ridiculous',
+  'Quirky',
+  'Playful',
+  'Amusing',
+  'Entertaining',
+  'Ludicrous',
+  'Absurd',
+  'Comical',
+  'Outlandish',
+  'Foolish',
+  'Wacky',
+  'Funky',
+  'Zesty',
+  'Jolly',
+  'Delightful',
+  'Krombopulos',
+  'Flingus',
+  'Zogon',
+  'Flurb',
+  'Jabberwocky',
+  'Snickerdoodle',
+  'Wobbegong',
+  'Mxyzptlk',
+  'Blunderbuss',
+  'Snollygoster',
+  'Quibblesnook',
+  'Noodlewhack',
+  'Flummox',
+  'Zizzwizz',
+  'Bumbershoot',
+  'Cantankerous',
+  'Scrumdiddlyumptious',
+];
+const LAST_NAME: string[] = [
+  'Banana',
+  'Penguin',
+  'Jellyfish',
+  'Noodle',
+  'Marshmallow',
+  'Squid',
+  'Giraffe',
+  'Sausage',
+  'Pickles',
+  'Muffin',
+  'Toothpaste',
+  'Pancake',
+  'Watermelon',
+  'Bubblegum',
+  'Cupcake',
+  'Donut',
+  'Kangaroo',
+  'Cheeseburger',
+  'Pickle',
+  'Taco',
+  'Sushi',
+  'Cucumber',
+  'Pineapple',
+  'Tofu',
+  'Popsicle',
+  'Waffle',
+  'Potato',
+  'Avocado',
+  'Penguin',
+  'Koala',
+  'Banana',
+  'Marshmallow',
+  'Sausage',
+  'Noodle',
+  'Jellybean',
+  'Cupcake',
+  'Cauliflower',
+  'Squirrel',
+  'Panda',
+  'Gummybear',
+  'Giraffe',
+  'Lollipop',
+  'Pancake',
+  'Broccoli',
+  'Pickles',
+  'Squid',
+  'Cheeseburger',
+  'Snickerdoodle',
+  'Wobbegong',
+  'Zogon',
+  'Flurb',
+  'Jabberwocky',
+  'Snollygoster',
+  'Quibblesnook',
+  'Noodlewhack',
+  'Flummox',
+  'Zizzwizz',
+];
+
+export {
+  API_ISSUE_REPORT,
+  DEFAULT_USER,
+  BACKEND_URL,
+  ERROR_FAILED_COLLECTION,
+  ERROR_TOO_MANY,
+  MIDDLE_NAME,
+  LAST_NAME,
+};
